Split imgproxy endpoints into modules
